// src/data/types.ts
import type { Component } from 'vue';

export interface Sound {
  icon: Component | string;
  id: string;
  label: string;
  src: string;
}

export type Sounds = Array<Sound>;

export interface Category {
  icon: Component | string;
  id: string;
  sounds: Sounds;
  title: string;
}

export type Categories = Array<Category>;